+++ lib/plausible/stats/filters/where_builder.ex
@@ -72,35 +72,34 @@ defmodule Plausible.Stats.Filters.WhereBuilder do
     dynamic([e], e.name in ^list)
   end
 
-  defp add_filter(:events, _query, [:is, "event:goal", clauses]) do
-    {events, pages} = split_goals(clauses)
+  defp add_filter(:events, _query, [operation, "event:goal", clauses])
+       when operation in [:is, :matches] do
+    {events, pages, wildcard?} = split_goals(clauses)
 
-    dynamic([e], (e.pathname in ^pages and e.name == "pageview") or e.name in ^events)
-  end
+    if wildcard? do
+      event_clause =
+        if Enum.any?(events) do
+          dynamic([x], fragment("multiMatchAny(?, ?)", x.name, ^events))
+        else
+          dynamic([x], false)
+        end
 
-  defp add_filter(:events, _query, [:matches, "event:goal", clauses]) do
-    {events, pages} = split_goals(clauses, &page_regex/1)
+      page_clause =
+        if Enum.any?(pages) do
+          dynamic(
+            [x],
+            fragment("multiMatchAny(?, ?)", x.pathname, ^pages) and x.name == "pageview"
+          )
+        else
+          dynamic([x], false)
+        end
 
-    event_clause =
-      if Enum.any?(events) do
-        dynamic([x], fragment("multiMatchAny(?, ?)", x.name, ^events))
-      else
-        dynamic([x], false)
-      end
+      where_clause = dynamic([], ^event_clause or ^page_clause)
 
-    page_clause =
-      if Enum.any?(pages) do
-        dynamic(
-          [x],
-          fragment("multiMatchAny(?, ?)", x.pathname, ^pages) and x.name == "pageview"
-        )
-      else
-        dynamic([x], false)
-      end
-
-    where_clause = dynamic([], ^event_clause or ^page_clause)
-
-    dynamic([e], ^where_clause)
+      dynamic([e], ^where_clause)
+    else
+      dynamic([e], (e.pathname in ^pages and e.name == "pageview") or e.name in ^events)
+    end
   end
 
   defp add_filter(:events, _query, [_, "event:page" | _rest] = filter) do
@@ -189,6 +188,44 @@ defmodule Plausible.Stats.Filters.WhereBuilder do
     )
   end
 
+  defp filter_custom_prop(prop_name, column_name, [:does_not_match, _, clauses]) do
+    regexes = Enum.map(clauses, &page_regex/1)
+
+    dynamic(
+      [t],
+      has_key(t, column_name, ^prop_name) and
+        fragment(
+          "not(arrayExists(k -> match(?, k), ?))",
+          get_by_key(t, column_name, ^prop_name),
+          ^regexes
+        )
+    )
+  end
+
+  defp filter_custom_prop(prop_name, column_name, [:contains, _, clauses]) do
+    dynamic(
+      [t],
+      has_key(t, column_name, ^prop_name) and
+        fragment(
+          "multiSearchAny(?, ?)",
+          get_by_key(t, column_name, ^prop_name),
+          ^clauses
+        )
+    )
+  end
+
+  defp filter_custom_prop(prop_name, column_name, [:does_not_contain, _, clauses]) do
+    dynamic(
+      [t],
+      has_key(t, column_name, ^prop_name) and
+        fragment(
+          "not(multiSearchAny(?, ?))",
+          get_by_key(t, column_name, ^prop_name),
+          ^clauses
+        )
+    )
+  end
+
   defp filter_field(db_field, [:matches, _key, glob_exprs]) do
     page_regexes = Enum.map(glob_exprs, &page_regex/1)
     dynamic([x], fragment("multiMatchAny(?, ?)", field(x, ^db_field), ^page_regexes))
@@ -197,6 +234,14 @@ defmodule Plausible.Stats.Filters.WhereBuilder do
   defp filter_field(db_field, [:does_not_match, _key, glob_exprs]) do
     page_regexes = Enum.map(glob_exprs, &page_regex/1)
     dynamic([x], fragment("not(multiMatchAny(?, ?))", field(x, ^db_field), ^page_regexes))
+  end
+
+  defp filter_field(db_field, [:contains, _key, values]) do
+    dynamic([x], fragment("multiSearchAny(?, ?)", field(x, ^db_field), ^values))
+  end
+
+  defp filter_field(db_field, [:does_not_contain, _key, values]) do
+    dynamic([x], fragment("not(multiSearchAny(?, ?))", field(x, ^db_field), ^values))
   end
 
   defp filter_field(db_field, [:is, _key, list]) do
@@ -222,13 +267,14 @@ defmodule Plausible.Stats.Filters.WhereBuilder do
   defp db_field_val(_, @not_set), do: ""
   defp db_field_val(_, val), do: val
 
-  defp split_goals(clauses, map_fn \\ &Function.identity/1) do
-    groups =
-      Enum.group_by(clauses, fn {goal_type, _v} -> goal_type end, fn {_k, val} -> map_fn.(val) end)
+  defp split_goals(clauses) do
+    wildcard? = Enum.any?(clauses, fn {_, value} -> String.contains?(value, "*") end)
+    map_fn = if(wildcard?, do: &page_regex/1, else: &Function.identity/1)
 
-    {
-      Map.get(groups, :event, []),
-      Map.get(groups, :page, [])
-    }
+    clauses
+    |> Enum.reduce({[], [], wildcard?}, fn
+      {:event, value}, {event, page, wildcard?} -> {event ++ [map_fn.(value)], page, wildcard?}
+      {:page, value}, {event, page, wildcard?} -> {event, page ++ [map_fn.(value)], wildcard?}
+    end)
   end
 end